+++ NEWS.md
@@ -2,6 +2,8 @@
 ==================
 ### Bug Fixes
 - Fix (and simplify) balanceDF.plot to organize the order of groups (now unadjusted/self is left, adjusted/self center, and target is on the right)
+- Fix plotly functions to use the red color for self when only compared to target (since in that case it is likely unadjusted): balance.stats_and_plots.weighted_comparisons_plots.plotly_plot_qq and balance.stats_and_plots.weighted_comparisons_plots.plotly_plot_bar
+
 
 0.1.0 (2022-11-20)
 ==================